# swagger.json
{
  "swagger": "2.0",
  "info": {
      "description": "Access and Modify MARE Organizational Data Via the API ",
      "version": "1.0.1",
      "title": "MARE API",
      "contact": {
          "email": "isaac@chicosystems.com"
      },
      "license": {
          "name": "Private License",
          "url": "https://maregroup.org"
      }
  }, 
  "schemes": ["http"],
  "host": "127.0.0.1:3081",
  "basePath": "/api",
  "paths" : {
        "/session" : {
            "post" : {
                "summary" : "Record A New Session",
                "description": "Record A New Session into the MARE Session Database.",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Session Array",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "properties": {
                                "session" : {
                                    "type": "array",
                                    "$ref": "#/definitions/session"
                                } 
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/session"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/sessions" : {
            "get" : {
                "summary" : "Get All The Sessions In The DB",
                "description": "This returns a JSON Object Representing All The Sessions in the Database. Warning! If the Database is too large, this can take a lot of time to run.",
                "produces": ["application/json"],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/session"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Status Value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/observation/getLastVideoInfo/{session_id}" : {
            "get" : {
                "summary" : "Gets the Video Info for the max observation in the given session.",
                "description": "This returns a JSON Object Representing All The Sessions in the Database. Warning! If the Database is too large, this can take a lot of time to run.",
                "produces": ["application/json"],
                "parameters": [ {
                    "name": "session_id",
                    "in": "path",
                    "description": "The ID of the session for which we are getting the LAST video",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/session"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Status Value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/observation/updateObservationWithCount/{session_id}/{obsID}/{count}" : {
            "get" : {
                "summary" : "Updates the given observation_id with a new count.",
                "description": "Returns 1 if a count was updated, 0 if no count was updated..",
                "produces": ["application/json"],
                "parameters": [ 
                {
                    "name": "session_id",
                    "in": "path",
                    "description": "The session_id that contains the obsID that we are updating..",
                    "required": true,
                    "type": "string"
                },{
                    "name": "obsID",
                    "in": "path",
                    "description": "The ID of the observation for which we are updating the count.",
                    "required": true,
                    "type": "string"
                },
                {
                    "name": "count",
                    "in": "path",
                    "description": "The new count that we are updating.",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/session"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Status Value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },"/session/" : {
            "put" : {
                "summary" : "Update A Given Session By ID",
                "description": "Updates An Session Identified by its ID. Takes the body, and updates any values passed in. ",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Object to be Updated",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Session Array",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "properties": {
                                "session" : {
                                    "type": "array",
                                    "$ref": "#/definitions/session"
                                } 
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/session"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Status Value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/session/{id}" : {
            "delete" : {
                "summary" : "Delete a Given Session by ID",
                "description": "Delete an Session when given its id.",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Session id that needs to be deleted",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/todosResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/session/createNewSession/{processorName}/{projectName}/{line}/{dive}/{lineID}/{type}" : {
            "post" : {
                "summary" : "Creates a new session if it doesn't exist. Also create new processor name, and/or project Name, if they do not exist.",
                "description": "Create New session, and possiblly new processor name and project name.",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "processorName",
                        "in": "path",
                        "description": "The Name of the processor we are creating a new session for. If name doesn't exist it will be created.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "projectName",
                        "in": "path",
                        "description": "The Name of the Project we are creating a new session for. If the name doesn't exist we will create it.",
                        "required": true,
                        "type": "string"
                    }
                    ,
                    {
                        "name": "line",
                        "in": "path",
                        "description": "The Line Number we are creating a session for..",
                        "required": true,
                        "type": "string"
                    },{
                        "name": "dive",
                        "in": "path",
                        "description": "The Dive we are creating a session for..",
                        "required": true,
                        "type": "string"
                    },{
                        "name": "lineID",
                        "in": "path",
                        "description": "The lineID we are creating a session for..",
                        "required": true,
                        "type": "string"
                    },{
                        "name": "type",
                        "in": "path",
                        "description": "The type we are creating a session for..",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/session"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/sessions/user/{userID}/project/{projectID}" : {
            "get" : {
                "summary" : "Gets All The Sessions That are Associated with a Specific User, and a Specific Project.",
                "description": "This returns a JSON Object Representing All The Sessions associated with the given user, on the given project.",
                "produces": ["application/json"],
                
                "parameters": [
                    {
                        "name": "userID",
                        "in": "path",
                        "description": "The ID of the User",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "projectID",
                        "in": "path",
                        "description": "The ID of the Project",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/session"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Status Value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/project" : {
            "post" : {
                "summary" : "Record A New Project",
                "description": "Record A New Project into the MARE Database.",
                "produces": ["application/json"],
                "consumes": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Project Array",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "properties": {
                                "project" : {
                                    "type": "array",
                                    "$ref": "#/definitions/project"
                                } 
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/project"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid status value",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
...
  },
  "definitions": {
      "todosResponse": {
          "type": "object",
          "properties": {
              "_id": {
                   "type": "integer"
              },
              "name": {
                  "type": "string"
              },
              "description": {
                  "type": "string"
              },
              "createdBy": {
                  "type": "string"
              },
              "createdAt": {
                  "type": "string"
              },
              "updatedAt": {
                  "type": "string"
              },
              "updatedBy": {
                  "type": "string"
              }
          }
      },
...
    "session": {
        "type": "object",
        "properties": {
            "session_id": {
                 "type": "integer"
            },
            "user_id": {
                "type": "integer"
            },
            "project_id": {
                "type": "integer"
            },
            "dive": {
                "type": "string"
            },
            "line": {
                "type": "string"
            },
            "lineId": {
                "type": "string"
            },
            "type": {
                "type": "string"
            }
        }
      },
...
      "InvalidResponse": {
          "type": "object",
          "properties": {
              "statusCode": {
                  "type": "string"
              },
              "message": {
                  "type": "string"
              }
          }

      }
  }
}